matlab_script works fine under Windows 7, for both .jar and eclipse project

+++ matlab_script/agent_terminalStep.m
@@ -8,8 +8,8 @@ import edu.cwru.SimpleRTS.model.*;
 import edu.cwru.SimpleRTS.model.resource.*;
 import edu.cwru.SimpleRTS.model.unit.*;
 
-currentGold = state.getResourceAmount(0, ResourceType.GOLD);
-currentWood = state.getResourceAmount(0, ResourceType.WOOD);
+currentGold = state.getResourceAmount(0, ResourceType.GOLD).intValue();
+currentWood = state.getResourceAmount(0, ResourceType.WOOD).intValue();
 fprintf('\nCurrent Gold: %d\n', currentGold);
 fprintf('Current Wood: %d\n', currentWood);
 fprintf('Congratulations! You finish the task');
